+++ tests/test_unity_stac.py
@@ -26,30 +26,34 @@ def test_read_stac():
     # Added 8/10/23 to check the STAC collection information
     assert datasets[1].collection_id == 'C2011289787-GES_DISC'
 
-    data_files = collection.data_locations()
+    data_files = collection.data_files()
     assert len(data_files) == 6
-    data_files = collection.data_locations(["data","opendap"])
+    data_files = collection.data_files(["data","opendap"])
+    assert len(data_files) == 2
+    data_files = collection.data_files(["data","opendap","metadata"])
     assert len(data_files) == 4
-    data_files = collection.data_locations(["data","opendap","metadata"])
-    assert len(data_files) == 6
     data_files = collection.data_locations(["data"])
     assert len(data_files) == 2
-    for x in data_files:
+    data_locations = collection.data_locations(["data"])
+    for x in data_locations:
         assert x in ['https://data.gesdisc.earthdata.nasa.gov/data/CHIRP/SNDR13CHRP1.2/2016/235/SNDR.SS1330.CHIRP.20160822T0005.m06.g001.L1_AQ.std.v02_48.G.200425095850.nc', 'https://data.gesdisc.earthdata.nasa.gov/data/CHIRP/SNDR13CHRP1.2/2016/235/SNDR.SS1330.CHIRP.20160822T0011.m06.g002.L1_AQ.std.v02_48.G.200425095901.nc']
 
-        #Try a "classic" catalog + item files stac catalog
+    #Try a "classic" catalog + item files stac catalog
     collection = Collection.from_stac("tests/test_files/catalog_01.json")
     datasets = collection.datasets
     # Added 8/10/23 to check the STAC collection information
     assert datasets[0].collection_id == 'collection_test'
     assert len(datasets) == 1
-    data_files = collection.data_locations()
+    data_files = collection.data_files()
     assert len(data_files) == 2
-    data_files = collection.data_locations(["data"])
+    data_files = collection.data_files(["data", "metadata"])
+    assert len(data_files) == 2
+    data_files = collection.data_files(["data"])
     assert len(data_files) == 1
-    data_files = collection.data_locations(["metadata_stac"])
+    assert data_files[0].roles == ["data"]
+    data_files = collection.data_files(["metadata"])
     assert len(data_files) == 1
-    assert data_files[0] == "/unity/ads/sounder_sips/chirp_test_data/SNDR.SS1330.CHIRP.20160829T2317.m06.g233.L1_AQ.std.v02_48.G.200425130422.json"
+    assert data_files[0].roles == ["metadata"]
 
 
 def test_write_stac():